trying to fix error message for MAC address
The driver was taken from Linux v5.10 and is used now for version 6.5.
Apparently it is missing some API call(s) to prevent "soft" error message(s).
In this case the error message is harmless, and has no impact on the functioning of the device, but still a fix is being pursued.